refactor(content): move media library admin URL generation into Admin
Replace Content's MediaCrudController dependency with a Content contract
implemented by EasyAdmin in the Admin context.
