+++ src/Content/Application/Contract/MediaLibraryAdminUrlProviderInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Content\Application\Contract;
+
+interface MediaLibraryAdminUrlProviderInterface
+{
+    public function getIndexUrl(): string;
+}
